cleanup already moved files and dependencies
