+++ lib/config-loader.js
@@ -13,9 +13,15 @@ governing permissions and limitations under the License.
 const path = require('path')
 const yaml = require('js-yaml')
 const fs = require('fs')
-const utils = require('./utils')
 const aioConfig = require('@adobe/aio-lib-core-config')
 const aioLogger = require('@adobe/aio-lib-core-logging')('@adobe/aio-app-scripts:config-loader', { provider: 'debug' })
+
+function checkFile (filePath) {
+  // note lstatSync will throw if file doesn't exist
+  if (!fs.lstatSync(filePath).isFile()) {
+    throw Error(`${filePath} is not a valid file (e.g. cannot be a dir or a symlink)`)
+  }
+}
 
 // defaults
 const defaultAioHostname = 'adobeio-static.net'
@@ -124,7 +130,7 @@ module.exports = () => {
 
   // 2. check needed files
   aioLogger.debug('checking package.json existence')
-  utils.checkFile(_abs('package.json'))
+  checkFile(_abs('package.json'))
 
   // 3. load app config from package.json
   const packagejson = JSON.parse(fs.readFileSync(_abs('package.json')))
@@ -137,9 +143,9 @@ module.exports = () => {
     config.manifest.packagePlaceholder = '__APP_PACKAGE__'
     config.manifest.full = yaml.safeLoad(fs.readFileSync(config.manifest.src, 'utf8'))
     config.manifest.package = config.manifest.full.packages[config.manifest.packagePlaceholder]
-    if (config.manifest.package) {
-      aioLogger.debug(`Use of ${config.manifest.packagePlaceholder} in manifest.yml.`)
-    }
+    // if (config.manifest.package) {
+    //   aioLogger.debug(`Use of ${config.manifest.packagePlaceholder} in manifest.yml.`)
+    // }
     // Note: we should probably set the config.manifest.package also if it's not using a placeholder
   }
