+++ package.json
@@ -31,30 +31,18 @@
     "@adobe/aio-lib-core-config": "^2.0.0",
     "@adobe/aio-lib-core-logging": "^1.1.0",
     "@adobe/aio-lib-core-tvm": "^1.0.3",
-    "@adobe/aio-lib-runtime": "^0.1.0",
     "@hapi/joi": "^16.1.7",
-    "archiver": "^3.1.1",
     "aws-sdk": "^2.600.0",
-    "chokidar": "^3.3.1",
-    "debug": "^4.1.1",
-    "dotenv": "^8.2.0",
-    "execa": "^4.0.0",
     "fs-extra": "^9.0.0",
-    "globby": "^11.0.1",
-    "http-terminator": "^2.0.2",
     "js-yaml": "^3.13.1",
     "klaw": "^3.0.0",
     "lodash.clonedeep": "^4.5.0",
     "mime-types": "^2.1.24",
-    "node-fetch": "^2.6.0",
-    "open": "^7.0.0",
-    "parcel-bundler": "^1.12.4",
-    "webpack": "^4.41.5"
+    "parcel-bundler": "^1.12.4"
   },
   "devDependencies": {
     "@types/hapi__joi": "^16.0.2",
     "@types/jest": "^25.1.0",
-    "eol": "^0.9.1",
     "eslint": "^6.5.1",
     "eslint-config-standard": "^14.1.0",
     "eslint-plugin-import": "^2.18.2",
@@ -63,7 +51,6 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
     "jest": "^24.9.0",
-    "memfs": "^3.0.1",
-    "stdout-stderr": "^0.1.9"
+    "memfs": "^3.0.1"
   }
 }
